Move statics to wagtail_tag_manager folder, fix statics
This was causing issues with ManifestStaticFilesStorage raising a
ValueError because it could not find the statics of this library.

Django forms.Media expects them to just be plain paths: https://docs.djangoproject.com/en/3.0/topics/forms/media/

Furthermore, it was placing them in the static folder directly, which
could theoretically cause name clashes?).